refactor(xtts): remove duplicate xtts audio config
